+++ .gitignore
@@ -1,2 +1,4 @@
 #.env
 node_modules/.ngrok-irUW3I4o/bin/ngrok.exe
+frontend/src/server.js
+frontend/src/server.js
