+++ frontend/.gitignore
@@ -18,6 +18,9 @@
 .env.test.local
 .env.production.local
 
+#
+./frontend/server.js
+
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
