+++ frontend/src/server.js
@@ -1,5 +1,5 @@
-//export const server ="http://localhost:8000/api/v2";
-export const server ="https://backend-topaz-ten.vercel.app/api/v2";
+export const server ="http://localhost:8000/api/v2";
+//export const server ="https://backend-topaz-ten.vercel.app/api/v2";
 
-//export const backend_url = "http://localhost:8000/"
-export const backend_url = "https://backend-topaz-ten.vercel.app/"
+export const backend_url = "http://localhost:8000/"
+//export const backend_url = "https://backend-topaz-ten.vercel.app/"
